+++ frontend/packages/pipelines-plugin/src/components/pipelines/modals/common/SecretForm.tsx
@@ -9,7 +9,7 @@ import {
   SSHAuthSubform,
   PullSecretForm,
   SecretStringData,
-  SecretChangeData,
+  OnSecretChange,
 } from '@console/internal/components/secrets/create-secret';
 import { ButtonBar } from '@console/internal/components/utils';
 import { DropdownField, InputField, ActionGroupWithIcons } from '@console/shared';
@@ -19,7 +19,7 @@ import './SecretForm.scss';
 const renderSecretForm = (
   type: SecretType,
   stringData: { [key: string]: SecretStringData },
-  onDataChanged: (value: SecretChangeData) => void,
+  onDataChanged: OnSecretChange,
 ) => {
   switch (type) {
     case SecretType.basicAuth:
@@ -134,7 +134,7 @@ const SecretForm: React.FC<FormikProps<SecretFormValues>> = ({
     }
   };
 
-  const onDataChanged = (value: SecretChangeData) => {
+  const onDataChanged: OnSecretChange = (value) => {
     setStringData((prevState) => {
       setValues(values.type, { ...prevState, [values.type]: value });
       return { ...prevState, [values.type]: value };
