+++ frontend/public/components/secrets/create-secret/SecretSubForm.tsx
@@ -1,6 +1,9 @@
 import * as React from 'react';
-import { SourceSecretForm, PullSecretForm, WebHookSecretForm, GenericSecretForm } from '.';
 import { SecretSubFormProps, SecretTypeAbstraction } from './types';
+import { SourceSecretForm } from './SourceSecretForm';
+import { PullSecretForm } from './PullSecretForm';
+import { WebHookSecretForm } from './WebHookSecretForm';
+import { GenericSecretForm } from './GenericSecretForm';
 
 export const SecretSubForm: React.FC<
   SecretSubFormProps & { typeAbstraction: SecretTypeAbstraction }
