+++ frontend/public/components/secrets/create-secret/const.ts
@@ -1,1 +1,2 @@
 export const AUTHS_KEY = 'auths';
+export const WebHookSecretKey = 'WebHookSecretKey';
