+++ frontend/public/components/secrets/create-secret/utils.ts
@@ -1,9 +1,8 @@
 import * as _ from 'lodash-es';
-import { WebHookSecretKey } from '../../secret';
+import { WebHookSecretKey } from './const';
 import { useTranslation } from 'react-i18next';
-import { SecretTypeAbstraction, SecretType } from './types';
+import { SecretTypeAbstraction, SecretType, PullSecretCredential } from './types';
 import { Base64 } from 'js-base64';
-import { PullSecretCredential } from './PullSecretCredentialsForm';
 
 export const toDefaultSecretType = (typeAbstraction: SecretTypeAbstraction): SecretType => {
   switch (typeAbstraction) {
